requirements.txt
streamlit>=1.30.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
chromadb>=0.4.22
pdfplumber>=0.10.3
camelot-py>=0.11.0
opencv-python>=4.8.0.76
python-dotenv>=1.0.0
tiktoken>=0.5.2
